+++ src/main/java/subway/domain/Print.java
@@ -1,0 +1,14 @@
+package subway.domain;
+
+public class Print {
+    public static void mainMenu() {
+        System.out.printf(Constant.HASH + Constant.MAIN_MENU + "%n"
+                + Constant.ONE + Constant.DOT + Constant.STATION + Constant.SETTING + "%n"
+                + Constant.TWO + Constant.DOT + Constant.LINE + Constant.SETTING + "%n"
+                + Constant.THREE + Constant.DOT + Constant.SECTION + Constant.SETTING + "%n"
+                + Constant.FOUR + Constant.DOT + Constant.PRINT_SUBWAY_MAP + "%n"
+                + Constant.Q + Constant.DOT + Constant.QUIT + "%n%n"
+                + Constant.HASH + Constant.CHOOSE_FUNCTION
+        );
+    }
+}
